admin/requirements.txt
streamlit
pypdf
langchain>=0.2.6
langchain-community>=0.2.6
faiss-cpu
boto3
python-dotenv